# Speedup testCMCSingleMuscle