--- a/Applications/CMC/test/block_hanging_from_muscle_InitialStates.sto
+++ b/Applications/CMC/test/block_hanging_from_muscle_InitialStates.sto
@@ -1,6 +1,0 @@
-initial_states				
-nRows=1				
-nColumns=5				
-endheader				
-time	block_ty	block_ty_u	muscle.activation	muscle.fiber_length
-0	-0.42462006920006	0	0.02	0.122100534303908