… Applications/CMC/test/block_hanging_from_muscle.osim
… Applications/CMC/test/block_hanging_from_muscle_Setup_Forward.xml
… Applications/CMC/test/block_hanging_from_muscle_controls.sto
… Applications/CMC/test/block_hanging_from_muscle_controls.xml