Shorten and improve single muscle CMC test, which compares the results from CMC when tracking kinematics of a forward simulation using the same model. The model file version was updated to 3.0 (since test developed in 3.3) and the duration was shortened (5s->0.9s) and the peek excitation was doubled. Movement is more dynamic (faster and more range) and the controls and output forces are now being compared as well as states. 

In addition to being a harder test, this should also be at least 5x faster than the previous test case.
On Appveyor 32bit: [~180s](https://ci.appveyor.com/project/opensim-org/opensim-core/build/4152/job/33b7cof263yxc7ss#L495) to [~14s](https://ci.appveyor.com/project/opensim-org/opensim-core/build/4151/job/qhlgsm8780ko76p6#L473)