@tkuchida 2016-07-29
:+1: LGTM provided AppVeyor passes on 3e02d6b. This is a great way to speed up the tests (some were probably designed years ago when we relied on nightly builds, where speed wasn't a concern). However, I think tests like this should use models generated programmatically rather than storing and maintaining model files; separate tests should exercise model file loading and compare rigorously to ensure all properties are read in correctly, for example.