@aseth1 2016-08-01
>I think tests like this should use models generated programmatically rather than storing and maintaining model files; separate tests should exercise model file loading and compare rigorously to ensure all properties are read in correctly

I agree.  In this case building the model (programmatically) would be straightforward but setting up the CMCTool could be troublesome, since the Tools were never designed to be called directly from the API. I would prefer not to convert this test, but require that future tests (e.g. for a replacement for CMC and new solvers) have separate tests for performance, implementation correctness and de/serialization. 

Moving forward with new components, testing serialization, cloning, connectors and availability of outputs, can be tested by a generic set of cases as part of `testComponents`. So we can continue to reduce the dependency on setup and model files for testing core functionality as more functionality is encompassed by Components. As for Tools, I think it is just easier to include the setup files until we can deprecate them altogether (after 4.0), but that could also be a long time.

I could also attempt to setup the Forward and CMC Tools programmatically for this test. What is your preference for this PR?