@tkuchida 2016-08-01
> What is your preference for this PR?

No need to hold up this PR. I suggest opening a new issue for (1) generating models (and setup files, wherever possible) programmatically rather than loading them from xml, and (2) creating a new test focusing on serialization/deserialization that includes loading models saved in each supported version.