@aseth1 2016-08-01
Sounds good.

> (2) creating a new test focusing on serialization/deserialization that includes loading models saved in each supported version.

There are existing tests, in particular, `testSerializeOpenSimObjects` and `testComponents` that test serialization and deserialization and equality of models. However, `testSerializeOpenSimObjects` appears to be limited to `ModelComponents` which makes it far from being complete for testing OpenSim "Objects". I added an issue #1150 about this. There is an open issue for model conversion testing, #206.

> (1) generating models (and setup files, wherever possible) programmatically rather than loading them from xml

I opened an issue, #1151,  to propose adding a new guideline to our CONTRIBUTING.md to avoid the use of files (model or otherwise) for unit and functional testing (unless of course its function is to read/write data) and promote the use of short integration and regression tests.

If  this is satisfactory please feel free to merge.