@tkuchida 2016-08-01
> If this is satisfactory please feel free to merge.

LGTM. This PR exceeds what is covered by [Rule 5](https://github.com/opensim-org/opensim-core/blob/master/CONTRIBUTING.md#making-a-pull-request-pr); by Rule 6, a second reviewer needs to merge.