@klshrinidhi 2016-08-02
If possible, it would be good to not add any more code that uses `OpenSim::Array` and use `std::vector` instead.

Other than that, I don't see any problem. :+1: