@aseth1 2016-08-02
Thanks @klshrinidhi. I agree with your request. The added  `OpenSim::Array`s are used to define the tolerances as expected by `CHECK_STORAGE_AGAINST_STANDARD()` and updating that method would impact numerous test cases. Making that update perhaps should wait until we are closer to eliminating OpenSim::Array altogether. Otherwise, if we decide to back OpenSim::Array with std::vector, this might not be an issue either. A good discussion perhaps for a future PR?