@tkuchida 2016-08-18
> A good discussion perhaps for a future PR?

Have you created an issue?